+++ xnlib/xnacrylic/source/psf/psf.cpp
@@ -19,12 +19,12 @@
 #include <canvas.h>
 
 
-uint64_t *font_data = nullptr;
+uint8_t *font_data = nullptr;
 
 
 void psf_register_font_lib (){
 	valloc(0xFFFFFFFFA0000000);
-	font_data = (uint64_t*)0xFFFFFFFFA0000000;
+	font_data = (uint8_t*)0xFFFFFFFFA0000000;
 	UFILE file;
 	int fd = sys_open_file ("/font.psf", &file);
 	sys_read_file (fd,font_data,&file);
